# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: Backend API for the Heracle personal training & fitness platform.
  title: Heracle Fitness API
  version: "1.0"
paths:
...
  /sleep/:
    get:
      description: Returns the last 7 sleep data entries for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Sleep data
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get sleep data
      tags:
      - Sleep
    post:
      consumes:
      - application/json
      description: Adds a new sleep data entry for the authenticated user (keeps last
        7 entries)
      parameters:
      - description: Sleep data entry
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Updated sleep record
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add sleep data
      tags:
      - Sleep
  /sleep/insight:
    get:
      description: Returns an AI-generated insight based on recent sleep data (cached
        daily)
      produces:
      - application/json
      responses:
        "200":
          description: AI insight
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get AI sleep insight
      tags:
      - Sleep
  /split/:
    get:
      description: Returns the workout split assigned to the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Workout split
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Split not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get my workout split
      tags:
      - Split
  /split/trainer/{clientId}:
    get:
      description: Returns the workout split for a trainer's assigned client
      parameters:
      - description: Client user ID
        in: path
        name: clientId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Client workout split
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Not authorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Split not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Trainer: get client workout split'
      tags:
      - Split
    put:
      consumes:
      - application/json
      description: Creates or updates the workout split for a trainer's assigned client
      parameters:
      - description: Client user ID
        in: path
        name: clientId
        required: true
        type: string
      - description: Split data
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Upserted workout split
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Not authorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Trainer: create or update client workout split'
      tags:
      - Split
  /trainer/admin/add:
    post:
      consumes:
      - application/json
      description: Promotes an existing user to trainer role (admin only)
      parameters:
      - description: Trainer details
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Created trainer with user info
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Already a trainer
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Admin: add a trainer'
      tags:
      - Trainer
  /trainer/client/{clientId}:
    get:
      description: Returns detailed profile, metrics, and progress for a specific
        client
      parameters:
      - description: Client user ID
        in: path
        name: clientId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Client details
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Not authorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Client not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get client details
      tags:
      - Trainer
  /trainer/clients:
    get:
      description: Returns all clients assigned to the trainer with daily progress
      produces:
      - application/json
      responses:
        "200":
          description: Client list with progress
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "403":
          description: Trainer not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get trainer's clients
      tags:
      - Trainer
  /trainer/clients/add:
    post:
      consumes:
      - application/json
      description: Assigns a user as a client to the trainer by email
      parameters:
      - description: Client email
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Added client with progress
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Trainer not found
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Already assigned
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add a client
      tags:
      - Trainer
  /trainer/clients/remove/{clientId}:
    delete:
      description: Removes a client assignment from the trainer
      parameters:
      - description: Client user ID
        in: path
        name: clientId
        required: true
        type: string
      responses:
        "200":
          description: Removed
        "403":
          description: Not authorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove a client
      tags:
      - Trainer
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"